subresource-integrity/subresource-integrity.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Subresource Integrity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>

<div id="container"></div>
<script>
    var xorigin_anon_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-anon-script.js';

    var xorigin_creds_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-creds-script.js';

    var xorigin_ineligible_script = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-ineligible-script.js';

    var SRIScriptTest = function(pass, name, src, integrityValue, crossoriginValue) {
        this.pass = pass;
        this.name = "Script: " + name;
        this.src = src;
        this.integrityValue = integrityValue;
        this.crossoriginValue = crossoriginValue;
    }

    SRIScriptTest.prototype.execute = function() {
        var test = async_test(this.name);
        var e = document.createElement("script");
        e.src = this.src;
        e.setAttribute("integrity", this.integrityValue);
        if(this.crossoriginValue) {
            e.setAttribute("crossorigin", this.crossoriginValue);
        }
        if(this.pass) {
            e.addEventListener("load", function() {test.done()});
            e.addEventListener("error", function() {
                test.step(function(){ assert_unreached("Good load fired error handler.") })
            });
        } else {
           e.addEventListener("load", function() {
                test.step(function() { assert_unreached("Bad load succeeded.") })
            });
           e.addEventListener("error", function() {test.done()});
        }
        document.body.appendChild(e);
    };

    var xorigin_anon_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-anon-style.css';

    var xorigin_creds_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-creds-style.css';

    var xorigin_ineligible_style = location.protocol
      + '//www1.' + location.hostname + ':' + location.port
      + '/subresource-integrity/crossorigin-ineligible-style.css';

    // Style tests must be done synchronously because they rely on the presence
    // and absence of global style. Thus, instead of executing them one at a
    // time, the style tests are implemented as a queue that builds up a list of
    // tests, and then executes them one at a time.
    var SRIStyleTest = function(queue, pass, name, href, integrityValue, crossoriginValue) {
        this.pass = pass;
        this.name = "Style: " + name;
        this.href = href;
        this.integrityValue = integrityValue;
        this.crossoriginValue = crossoriginValue;

        this.test = async_test(this.name);

        this.queue = queue;
        this.queue.push(this);
    }

    SRIStyleTest.prototype.execute = function() {
        var container = document.getElementById("container");
        while (container.hasChildNodes()) {
          container.removeChild(container.firstChild);
        }

        var test = this.test;

        var div = document.createElement("div");
        div.className = "testdiv";
        var e = document.createElement("link");
        e.rel = "stylesheet";
        e.href = this.href;
        e.setAttribute("integrity", this.integrityValue);
        if(this.crossoriginValue) {
            e.setAttribute("crossorigin", this.crossoriginValue);
        }
        if(this.pass) {
            e.addEventListener("load", function() {
                test.step(function() {
                    var background = window.getComputedStyle(div, null).getPropertyValue("background-color");
                    assert_equals(background, "rgb(255, 255, 0)");
                    test.done();
                });
            });
            e.addEventListener("error", function() {
                test.step(function(){ assert_unreached("Good load fired error handler.") })
            });
        } else {
            e.addEventListener("load", function() {
                 test.step(function() { assert_unreached("Bad load succeeded.") })
             });
            e.addEventListener("error", function() {
                test.step(function() {
                    var background = window.getComputedStyle(div, null).getPropertyValue("background-color");
                    assert_equals(background, "rgba(0, 0, 0, 0)");
                    test.done();
                });
            });
        }
        container.appendChild(div);
        container.appendChild(e);
    };

    new SRIScriptTest(
        true,
        "Script: Same-origin with correct hash.",
        "matching-digest.js",
        "sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E="
    ).execute();

    new SRIScriptTest(
        false,
        "Same-origin with incorrect hash.",
        "non-matching-digest.js",
        "sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    ).execute();

    new SRIScriptTest(
        true,
        "Same-origin with multiple sha256 hashes, including correct.",
        "matching-digest.js",
        "sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E= sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    ).execute();

    new SRIScriptTest(
        true,
        "Same-origin with multiple sha256 hashes, including unknown algorithm.",
        "matching-digest.js",
        "sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E= foo666-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    ).execute();

    // TODO may need to change these to hash-source prefix, e.g. 'sha256' instead of 'SHA-256'

    test(function() {assert_equals(getPrioritizedHashFunction('SHA-512', 'SHA-256'), 'SHA-512')},
        "SHA-512 preferred to SHA-256.");

    test(function() {assert_equals(getPrioritizedHashFunction('SHA-384', 'SHA-512'), 'SHA-512')},
        "SHA-512 preferred to SHA-384.");

    test(function() {assert_equals(getPrioritizedHashFunction('SHA-384', 'SHA-256'), 'SHA-384')},
        "SHA-384 preferred to SHA-256.");

    test(function() {assert_equals(getPrioritizedHashFunction('MD5', 'SHA-256'), 'SHA-256')},
        "SHA-256 preferred to MD5.");

    test(function() {assert_equals(getPrioritizedHashFunction('SHA-256', 'SHA-256'), '')},
        "getPrioritizedHashFunction('SHA-256', 'SHA-256') returns empty string");

    new SRIScriptTest(
        true,
        "Same-origin with sha256 mismatch, sha512 match",
        "matching-digest.js",
        "sha512-geByvIIRspbnUnwooKGNNCb39nvg+EW0O9hDScTXeo/9pVZztLSUYU3LNV6H0lZapo8bCJUpyPPLAzE9fDzpxg== sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    ).execute();

    new SRIScriptTest(
        false,
        "Same-origin with sha256 match, sha512 mismatch",
        "matching-digest.js",
        "sha512-deadbeefspbnUnwooKGNNCb39nvg+EW0O9hDScTXeo/9pVZztLSUYU3LNV6H0lZapo8bCJUpyPPLAzE9fDzpxg== sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E="
    ).execute();

    new SRIScriptTest(
        true,
        "<crossorigin='anonymous'> with correct hash, ACAO: *",
        xorigin_anon_script,
        "sha256-r6KBpvzfcSLlbFZCj1OACLxTxVck2TOrBTEdUbwz1yU=",
        "anonymous"
    ).execute();

    new SRIScriptTest(
        false,
        "<crossorigin='anonymous'> with incorrect hash, ACAO: *",
        xorigin_anon_script,
        "sha256-deadbeefcSLlbFZCj1OACLxTxVck2TOrBTEdUbwz1yU=",
        "anonymous"
    ).execute();

    new SRIScriptTest(
        true,
        "<crossorigin='use-credentials'> with correct hash, CORS-eligible",
        xorigin_creds_script,
        "sha256-sPICZvGN2S+pTRZgiw3DWrhC6JLDlt2zRyGpwH7unU8=",
        "use-credentials"
    ).execute();

    new SRIScriptTest(
        false,
        "<crossorigin='use-credentials'> with incorrect hash CORS-eligible",
        xorigin_creds_script,
        "sha256-deadbeef2S+pTRZgiw3DWrhC6JLDlt2zRyGpwH7unU8=",
        "use-credentials"
    ).execute();

    new SRIScriptTest(
        false,
        "<crossorigin='anonymous'> with CORS-ineligible resource",
        xorigin_ineligible_script,
        "sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E=",
        "anonymous"
    ).execute();

    new SRIScriptTest(
        false,
        "Resource with Refresh header",
        "refresh-header.js",
        "sha256-ieQAXii4cMmZFLxSRnxfZ1KSyzCjOb+N2rQ6OaVBWyM="
    ).execute();

    new SRIScriptTest(
        false,
        "Resource with WWW-Authenticate header",
        "www-authenticate-header.js",
        "sha256-ztNCkGU1fBB5II5wihGTbFb9F2TIMaHldkbnMlp7G/M="
    ).execute();

    new SRIScriptTest(
        true,
        "Same-origin with correct hash, options.",
        "matching-digest.js",
        "sha256-EKclUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E=?foo=bar?spam=eggs"
    ).execute();

    new SRIScriptTest(
        true,
        "Same-origin with unknown algorithm only.",
        "matching-digest.js",
        "foo666-foolUXH9SRRUv3FmL7bIEV0z2s3EvzHFxzHKCnfHT/E="
    ).execute();

    var style_tests = [];
    style_tests.execute = function() {
        if (this.length > 0) {
            this.shift().execute();
        }
    }
    add_result_callback(function(res) {
        if (res.name.startsWith("Style: ")) {
          style_tests.execute();
        }
    });

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with correct hash",
        "style.css",
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F-SPLT7YZk7gyCWUV4="
    );

    new SRIStyleTest(
        style_tests,
        false,
        "Same-origin with incorrect hash.",
        "style.css",
        "sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with multiple sha256 hashes, including correct.",
        "style.css",
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F-SPLT7YZk7gyCWUV4= sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with multiple sha256 hashes, including unknown algorithm.",
        "style.css",
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F-SPLT7YZk7gyCWUV4= foo666-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with sha256 mismatch, sha512 match",
        "style.css",
        "sha512-9wXDjd6Wq3H6nPAhI9zOvG7mJkUr03MTxaO+8ztTKnfJif42laL93Be/IF6YYZHHF4esitVYxiwpY2HSZX4l6w== sha256-deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdead"
    );

    new SRIStyleTest(
        style_tests,
        false,
        "Same-origin with sha256 match, sha512 mismatch",
        "style.css",
        "sha512-deadbeef9wXDjd6Wq3H6nPAhI9zOvG7mJkUr03MTxaO+8ztTKnfJif42laL93Be/IF6YYZHHF4esitVYxiwpY2== sha256-CzHgdJ7wOccM8L89n4bhcJMz3F-SPLT7YZk7gyCWUV4="
    );

    new SRIStyleTest(
        style_tests,
        true,
        "<crossorigin='anonymous'> with correct hash, ACAO: *",
        xorigin_anon_style,
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk7gyCWUV4=",
        "anonymous"
    );

    new SRIStyleTest(
        style_tests,
        false,
        "<crossorigin='anonymous'> with incorrect hash, ACAO: *",
        xorigin_anon_style,
        "sha256-deadbeefCzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk=",
        "anonymous"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "<crossorigin='use-credentials'> with correct hash, CORS-eligible",
        xorigin_creds_style,
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk7gyCWUV4=",
        "use-credentials"
    );

    new SRIStyleTest(
        style_tests,
        false,
        "<crossorigin='use-credentials'> with incorrect hash CORS-eligible",
        xorigin_creds_style,
        "sha256-deadbeefCzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk=",
        "use-credentials"
    );

    new SRIStyleTest(
        style_tests,
        false,
        "<crossorigin='anonymous'> with CORS-ineligible resource",
        xorigin_ineligible_style,
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk7gyCWUV4=",
        "anonymous"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with correct hash, options.",
        "style.css",
        "sha256-CzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk7gyCWUV4=?foo=bar?spam=eggs"
    );

    new SRIStyleTest(
        style_tests,
        true,
        "Same-origin with unknown algorithm only.",
        "style.css",
        "foo666-CzHgdJ7wOccM8L89n4bhcJMz3F+SPLT7YZk7gyCWUV4=?foo=bar?spam=eggs"
    );

    style_tests.execute();

</script>
<!--TODO check cache-poisoned resources, transfer-encoding, 3xx redirect
  to resource with matching hash, and cross-origin leakage test as in sec5.3 -->
